Move init to edge package
