+++ cmd/demo/main.go
@@ -14,7 +14,7 @@ func main() {
 	}
 	defer w.Destroy()
 	w.SetTitle("Minimal webview example")
-	w.SetSize(800, 600, webview2.HintNone)
+	w.SetSize(800, 600, webview2.HintFixed)
 	w.Navigate("https://en.m.wikipedia.org/wiki/Main_Page")
 	w.Run()
 }
